requirements.txt
selenium==4.27.1 
requests==2.32.3 
openai-whisper==20240930 
openpyxl==3.1.5 
ffmpeg-python==0.2.0 

